Add files via upload
This project was undertaken as part of my Bachelor's degree in Computer and Communication Engineering with a minor in Biomedical Engineering. The project involved designing a robotic arm with sensation capabilities, which could detect heat and pressure and respond automatically..
The project received full funding from LIRA (Lebanese Industrial Research Achievements), which was provided by the Ministry of Industry in Lebanon in 2018. The project was recognized as one of the top 10 projects of the year.
The significance of this project lies not only in the development of robots that can protect themselves from harm, but also in the potential for the technology to be used in the creation of prosthetic arms for amputees. The objective of the project was to design and develop a robotic arm that could sense temperature, pain, and pressure. To achieve this, multiple temperature and pressure sensors were placed on the palm and fingers of the robotic hand, and four servo motors were placed on the elbow to control the arm's movements. Pain sensation was obtained by carefully studying human temperature and pressure thresholds and embedding them in custom-developed software. The measured values were represented by a withdrawal reflex achieved by the servo motors to protect the arm from potential damage. Additionally, a voice signal depicted the resulting sensation, and a screen displayed the level of temperature, pain, or pressure, or any combination of them.
The arm was designed to have a fast response time, and multiple sensors were tested before their incorporation. In the future, this arm may be implemented on a robot, and as a first step, it may be further developed to function as a prosthetic arm for amputees.